requirements.txt
datapizza-ai==0.0.9
openai==2.9.0
# IMPORTANTE: Blocchiamo la versione alla 1.x per stabilità e compatibilità
e2b-code-interpreter==1.5.2
pandas==2.3.3
openpyxl==3.1.5
pyyaml==6.0.3
streamlit==1.52.1
python-dotenv==1.2.1
pydantic==2.12.5
# Fix SSL per Windows (fondamentale per alcuni ambienti corporate/locali)
pip-system-certs; sys_platform == 'win32'
